requirements.txt
Flask
Flask-Bootstrap
Flask-Login
Flask-Moment
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Mail
Jinja2
Werkzeug
WTForms
WTForms-SQLAlchemy
email-validator
pytest
python-dotenv
psycopg2
google
google-generativeai
ics
tensorflow
pillow
ultralytics
transformers
deep_translator
google-cloud-texttospeech